chore: move range validation into MCDARangeControls component
